pyproject.toml
[build-system]
requires = [
  "setuptools>=64.0",
  "setuptools-scm>=8.0",
]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
version_file = "cliapp/_version.py"

[tool.setuptools]
packages = ["cliapp", "simple"]
py-modules = ["__init__"]

[project]
name = "cliapp"
dynamic = ["version"]
authors = [
  { name = "Jaroslaw Wencel", email = "jaroslaw@wencel.it" },
]
maintainers = [
  { name = "Jaroslaw Wencel", email = "jaroslaw@wencel.it" },
]
description = "Placeholder"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4~=4.14.3",
  "dnspython~=2.8.0",
  "httpx~=0.28.1",
  "openpyxl~=3.1.5",
  "python-dotenv~=1.2.1",
  "ruamel-yaml~=0.19.1",
  "tabulate~=0.9.0",
  "tqdm~=4.67.1",
  "validators~=0.35.0",
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

[project.optional-dependencies]
dev = [
  "pylint~=4.0.4",
  "pip-check-updates~=0.28.0",
]

[project.urls]
Homepage = "https://serafin.tech/python-scripts"
Issues = "https://serafin.tech/python-scripts/issues"

[project.scripts]
cliapp = "cliapp:cliapp.main"
